First attempt at CDF for non-isotropic emission.
What is included is the working code for an initial implementation of the non-isotropic emission CDF / sampling code. This code uses bins in the CDF to represent the LID functions of light sources.

However, I have already spotted a number of issues with this approach, so I think I will re-engineer this shortly.