Improved sampling on Spherical CDF code.
The code will now check to see if the distribution is finely enough sampled to capture the sin term in the surface integral. If the threshold is not met, it will automatically sample the underlying distribution with an appropriate sampling.

Also, updated NetCDF library version so that we no longer have to manually use GitHub.