Re-implemented non-isotropic emission.
This implementation is more robust and uses the existing `crate::math::rng:: Probability' object to generate the CDF and perform the sampling. I have converted the spherical CDF object to use this for both azimuthal and polar CDFs. I have also implemented a basic CDF / PDF output function for the LinearSpline variant of the `Probability` object so that we can check the output.