Improved reliability of MCRT draws.
As part of implementing one of the unit tests, which is also included in this commit, I have patched an edge case for the case of zero gradient in the Probability struct. I have also completely re-implemented how planes are found in a spherical CDF. Both code is now more reliable for the case where planes are equally spaced.

Also accompanying this commit is an implementation of dot product for Vec2.