Added support for sampling from non-evenly spaced planes.
With the updated version of `lidrs`, I have now added support for non-evenly spaced planes in the spherical CDF object. This means that we are now much more generalised to photometric data. For now I am opting to use the `lidrs` PlaneWidth struct for the spherical CDF, but we can use an internal representation if required.