Corrected Phi Lambertian Draws
I have corrected and refined the draws that we are performing on the phi angle on the Lambertian scattering. In this commit I have implemented a draw from the CDF, instead of a uniform draw. This looks to be mostly correct, but still room for improvement.

Todo: implement automated tests for theta and phi draws.