Fix for surface transition issues.
This is a fix for an issue where the code would panic at a surface interface when inside an object. As it transpires, the z-component of the Dir3 was not being negated in the neg implementation, causing the ray · norm to be negative; resulting in a panic. Also, modified some of the `Imager` code to avoid potential issues.

Todo: Implement test cases for Neg implementation.