Finalised reflectance code.
In this commit I have corrected the unit tests for composite reflectance and done some code cleanup around the reflectance code and removed the fixed the warnings. Importantly, I have changed the diffuse-to-specular ratio argument / parameter around - as diffuse comes first in the enum, it makes sense to do it this way.

This should mean it is good to go.