Rewritten Non-isotropic emitter.
I have rewritten non-isotropic emitters, and changed how they work. They now emit as a single source, with direction information. By default, they are placed at the origin and can be translated and rotated by a 3D transform, as geometry is. This allows us to orient light sources however we wish in the future.