Initial implementation of photon collector attrib.
This includes the photon collector attribute iteself, all of the Save implementation code for the attribute, and all of the linker implementation to make it work in the MCRT binary.

Optimisations required, as saving is slow, and there is also likely a sizable overhead for appending to the photon vector.