Fixed spectral reflectance bug.
The logic in the spectral interpolation let through photons that were above the maximum wavelength in the loaded spectrum. It should not do this. So, I have fixed this with an extract check before performing interpolation. I have also written a new unit test to check this behaviour in the reflectance class.