Added potential fix for shift map indexing in `fluorescence` engine.
The removal of the `grid` parameter also removed access to a consistent
measurement grid over which to index into the shifts map in the fluorescence engine. However, I have engineered an alternate solution which indexes using
the bounds of the simulation and takes from the res of the grid, potentially
making this solution more flexible.