Slimmed down attribute linking protptype.
Instead of having layer upon layer of linker types that each handle a different register of outputs, I am experimenting with linking more broadly while remaining semantic. For example, included in this experiment, we link all the outputs in one step and all of the materials in another step. This is clearer and easier to maintain. All that is required is a way to keep the interface consistent for the `Link` trait, as I've just implemented link in the impl for the output linker, rather than inside the trait.